+++ src/Model/Reports/InvoicingRateReportFormData.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Model\Reports;
+
+class InvoicingRateReportFormData
+{
+    public WorkloadReportPeriodTypeEnum $viewPeriodType;
+    public int $year;
+    public bool $includeIssues;
+}