+++ src/Service/JiraApiService.php
@@ -16,10 +16,7 @@ use App\Model\Invoices\ProjectDataCollection;
 use App\Model\Invoices\VersionData;
 use App\Model\Invoices\WorklogData;
 use App\Model\Invoices\WorklogDataCollection;
-use App\Model\Planning\Issue;
 use App\Model\Planning\PlanningData;
-use App\Model\Planning\Project;
-use App\Model\Planning\Sprint;
 use App\Model\SprintReport\SprintReportData;
 use App\Model\SprintReport\SprintReportEpic;
 use App\Model\SprintReport\SprintReportIssue;
@@ -737,6 +734,8 @@ class JiraApiService implements DataProviderServiceInterface
 
                 $issueData->epicKey = $epicKey;
                 $issueData->epicName = $epicData->fields->summary ?? null;
+
+                $issueData->epics = null !== $issueData->epicName ? [$issueData->epicName] : [];
             }
 
             foreach ($fields->fixVersions ?? [] as $fixVersion) {