+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-182](https://github.com/itk-dev/economics/pull/182)
+  2597: Added invoicing rate report
+* [PR-185](https://github.com/itk-dev/economics/pull/186)
+  2597: Added epic migration command.
+* [PR-184](https://github.com/itk-dev/economics/pull/184)
+  3489: Workload report period averages.
+* [PR-183](https://github.com/itk-dev/economics/pull/183)
+  2597: Added epic relations.
+* [PR-187](https://github.com/itk-dev/economics/pull/187)
+  Updated symfony bundles.
 * [PR-189](https://github.com/itk-dev/economics/pull/189)
   Npm audit.
 * [PR-187](https://github.com/itk-dev/economics/pull/187)